Make smallcontainer an atom ui element
