+++ packages/component-dashboard/client/components/NewSubmitterScreen.js
@@ -1,27 +1,14 @@
 import React from 'react'
-import styled from 'styled-components'
 import { Box } from '@rebass/grid'
 import { TwoColumnLayout } from '@elifesciences/component-elife-ui/client/global'
-import media from '@elifesciences/component-elife-ui/client/global/layout/media'
 import {
   Paragraph,
   ImageWrapper,
   FooterPrivacy,
   NativeLink,
+  SmallCenterer,
 } from '@elifesciences/component-elife-ui/client/atoms'
 import { H1, Button } from '@pubsweet/ui'
-
-const SmallCenterer = styled(Box).attrs({
-  mx: 'auto',
-  mt: 5,
-  px: 5,
-  width: [1, 1, 1, 1000],
-})`
-  min-width: 0;
-  ${media.tabletPortraitUp`
-  margin-top: 0;
-`};
-`
 
 const NewSubmitterScreen = ({ createNewSubmission }) => (
   <SmallCenterer>
