+++ packages/component-elife-ui/client/atoms/index.js
@@ -19,6 +19,7 @@ export { default as Paragraph } from './Paragraph'
 export { default as PopOverPanel } from './PopOverPanel'
 export { default as RibbonAction } from './RibbonAction'
 export { default as SelectedItem } from './SelectedItem'
+export { default as SmallCenterer } from './SmallCenterer'
 export { default as StickyFooter } from './StickyFooter'
 export { default as Textarea } from './Textarea'
 export { default as TopBarContainer } from './TopBarContainer'
